Refine backend trait delegation
